refactor(autotest): Modify Advanced Search endpoint test to use new BaseEntitySummary
