+++ autotest/OldTests/src/test/java/io/github/openequella/rest/AdvancedSearchApiTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 
 import com.tle.webtests.framework.TestConfig;
 import com.tle.webtests.framework.TestInstitution;
+import io.github.openequella.rest.models.BaseEntitySummary;
 import java.io.IOException;
 import java.util.List;
 import org.apache.commons.httpclient.HttpMethod;
@@ -27,39 +28,17 @@ public class AdvancedSearchApiTest extends AbstractRestApiTest {
 
   @Test
   public void testRetrieveAdvancedSearches() throws Exception {
-    final List<AdvancedSearchSummary> searches = getAdvancedSearches();
+    final List<BaseEntitySummary> searches = getAdvancedSearches();
     assertEquals(
         "The number of returned advanced searches should match the institution total.",
         3,
         searches.size());
   }
 
-  private List<AdvancedSearchSummary> getAdvancedSearches() throws IOException {
+  private List<BaseEntitySummary> getAdvancedSearches() throws IOException {
     final HttpMethod method = new GetMethod(ADVANCEDSEARCH_API_ENDPOINT);
     assertEquals(HttpStatus.SC_OK, makeClientRequest(method));
     return mapper.readValue(
-        method.getResponseBodyAsString(), new TypeReference<List<AdvancedSearchSummary>>() {});
-  }
-
-  // Mirror of com.tle.web.api.settings.AdvancedSearchSummary
-  private static class AdvancedSearchSummary {
-    private String uuid;
-    private String name;
-
-    public String getUuid() {
-      return uuid;
-    }
-
-    public void setUuid(String uuid) {
-      this.uuid = uuid;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
+        method.getResponseBodyAsString(), new TypeReference<List<BaseEntitySummary>>() {});
   }
 }
