+++ autotest/OldTests/src/test/java/io/github/openequella/rest/models/BaseEntitySummary.java
@@ -1,0 +1,23 @@
+package io.github.openequella.rest.models;
+
+// Mirrored com.tle.web.api.entity.BaseEntitySummary
+public class BaseEntitySummary {
+  private String uuid;
+  private String name;
+
+  public String getUuid() {
+    return uuid;
+  }
+
+  public void setUuid(String uuid) {
+    this.uuid = uuid;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+}
